Extract reusable code to helper action
